eeprom function return type is u16

+++ include/gba/eeprom.h
@@ -13,9 +13,9 @@
  * @param unk_1 4 for 512 byte, 0x40 for 8k
  * @return 1 on invalid argument, 0 otherwise
  */
-u32 EEPROMConfigure(u16 unk_1);
-u32 EEPROMRead(u16 address, u16* data);
+u16 EEPROMConfigure(u16 unk_1);
+u16 EEPROMRead(u16 address, u16* data);
 u16 EEPROMCompare(u16 address, const u16* data);
-u32 EEPROMWrite0_8k_Check(u16 address, const u16* data);
+u16 EEPROMWrite0_8k_Check(u16 address, const u16* data);
 
 #endif // GUARD_GBA_EEPROM
